Github Issue #3: Improved auto scroll for undo/redo and text entry Github Issue #24: Merged some loops for faster execution, improved scroll percent precision

+++ mkedit.h
@@ -96,7 +96,7 @@ protected:
     void scrollValueUpdateHandle(int value);
 
 signals:
-    void scrollPercentUpdate(int percent);
+    void scrollPercentUpdate(double percent);
     void cursorPosChanged(bool hasSelection, int blockNumber );
     void fileSave();
     void enterKeyPressed(int blockNumber);
